Combine small source files into bigger ones
There's more that can be done later, but this is a good start
