+++ src/actors.cpp
@@ -1,12 +1,8 @@
 #include "actors.h"
-#include "actor_prototype_types.h"
 #include "game_rendering.h"
 #include "game_state.h"
-#include "room_types.h"
-#include "dungeon_types.h"
 #include "random.h"
 #include "asset_manager.h"
-#include "anim_types.h"
 #include "collision.h"
 #include <gtc/constants.hpp>
