+++ src/audio.cpp
@@ -1,5 +1,5 @@
 #include "audio.h"
-#include "audio_types.h"
+#include "core_types.h"
 #include <cstdlib>
 #include <SDL.h>
 #include <vector>
