--- src/bullet.h
@@ -1,8 +1,0 @@
-#pragma once
-#include "bullet_data.h"
-
-namespace Game {
-	extern const ActorInitFn bulletInitTable[BULLET_TYPE_COUNT];
-	extern const ActorUpdateFn bulletUpdateTable[BULLET_TYPE_COUNT];
-	extern const ActorDrawFn bulletDrawTable[BULLET_TYPE_COUNT];
-}
