+++ src/editor.cpp
@@ -1,5 +1,4 @@
 #include "editor.h"
-#include "editor_actor.h"
 #include "editor_assets.h"
 #include "rendering.h"
 #include "debug.h"
@@ -19,17 +18,12 @@
 #include "tilemap.h"
 #include "rendering_util.h"
 #include "game.h"
-#include "room_types.h"
 #include "game_rendering.h"
 #include "game_state.h"
 #include "actors.h"
-#include "actor_prototype_types.h"
 #include "audio.h"
 #include "collision.h"
 #include "random.h"
-#include "dungeon_types.h"
-#include "overworld_types.h"
-#include "anim_types.h"
 #include "asset_manager.h"
 #include "asset_serialization.h"
 #define GLM_ENABLE_EXPERIMENTAL
