+++ src/effect.cpp
@@ -1,6 +1,4 @@
-#include "effect.h"
 #include "actors.h"
-#include "actor_prototype_types.h"
 #include "game_rendering.h"
 #include "audio.h"
 #include <cstdio>
