--- src/effect.h
@@ -1,8 +1,0 @@
-#pragma once
-#include "effect_data.h"
-
-namespace Game {
-	extern const ActorInitFn effectInitTable[EFFECT_TYPE_COUNT];
-	extern const ActorUpdateFn effectUpdateTable[EFFECT_TYPE_COUNT];
-	extern const ActorDrawFn effectDrawTable[EFFECT_TYPE_COUNT];
-}
