+++ src/game_state.cpp
@@ -5,11 +5,7 @@
 #include "game_ui.h"
 #include "coroutines.h"
 #include "dialog.h"
-#include "room_types.h"
 #include "collision.h"
-#include "actor_prototype_types.h"
-#include "dungeon_types.h"
-#include "overworld_types.h"
 #include "asset_manager.h"
 #include "debug.h"
 #include "actors.h"
