--- src/pickup.h
@@ -1,8 +1,0 @@
-#pragma once
-#include "pickup_data.h"
-
-namespace Game {
-	extern const ActorInitFn pickupInitTable[PICKUP_TYPE_COUNT];
-	extern const ActorUpdateFn pickupUpdateTable[PICKUP_TYPE_COUNT];
-	extern const ActorDrawFn pickupDrawTable[PICKUP_TYPE_COUNT];
-}
