+++ src/player.cpp
@@ -1,14 +1,8 @@
-#include "player.h"
 #include "actors.h"
-#include "actor_prototype_types.h"
 #include "game_rendering.h"
 #include "game_input.h"
 #include "game_state.h"
 #include "dialog.h"
-#include "room_types.h"
-#include "dungeon_types.h"
-#include "overworld_types.h"
-#include "anim_types.h"
 #include "audio.h"
 #include "asset_manager.h"
 #include "tilemap.h"
