--- src/shared/room_types.h
@@ -1,48 +1,0 @@
-#pragma once
-#include "rendering_types.h"
-#include "tilemap_types.h"
-#include "actor_types.h"
-#include "memory_pool.h"
-#include "asset_types.h"
-
-enum RoomScreenExitDir : u8 {
-    SCREEN_EXIT_DIR_RIGHT,
-    SCREEN_EXIT_DIR_LEFT,
-    SCREEN_EXIT_DIR_BOTTOM,
-    SCREEN_EXIT_DIR_TOP,
-
-    SCREEN_EXIT_DIR_DEATH_WARP,
-};
-
-constexpr u32 ROOM_MAX_DIM_SCREENS = 4;
-constexpr u32 ROOM_MAX_SCREEN_COUNT = ROOM_MAX_DIM_SCREENS * ROOM_MAX_DIM_SCREENS;
-constexpr u32 ROOM_SCREEN_TILE_COUNT = VIEWPORT_WIDTH_METATILES * VIEWPORT_HEIGHT_METATILES;
-constexpr u32 ROOM_MAP_TILE_COUNT = ROOM_MAX_SCREEN_COUNT * 2;
-constexpr u32 ROOM_TILE_COUNT = ROOM_MAX_SCREEN_COUNT * ROOM_SCREEN_TILE_COUNT;
-
-struct RoomActor {
-    u32 id;
-    ActorPrototypeHandle prototypeHandle;
-    glm::vec2 position;
-};
-
-struct RoomTemplate {
-    u8 width;
-    u8 height;
-    u32 mapTileOffset;
-    Tilemap tilemap;
-    u32 actorCount;
-    u32 actorOffset;
-
-	inline u32 GetMapTileCount() const {
-		return width * height * 2; // 2 for each screen
-	}
-
-	inline BgTile* GetMapTiles() const {
-		return (BgTile*)((u8*)this + mapTileOffset);
-	}
-
-	inline RoomActor* GetActors() const {
-		return (RoomActor*)((u8*)this + actorOffset);
-	}
-};
