--- src/shared/tilemap_types.h
@@ -1,42 +1,0 @@
-#pragma once
-#include "rendering_types.h"
-#include "asset_types.h"
-#include <cassert>
-
-constexpr u32 TILESET_DIM = 16;
-constexpr u32 TILESET_DIM_LOG2 = 4;
-constexpr u32 TILESET_SIZE = TILESET_DIM * TILESET_DIM;
-constexpr u32 TILESET_DIM_ATTRIBUTES = TILESET_DIM >> 1;
-constexpr u32 TILESET_ATTRIBUTE_COUNT = TILESET_DIM_ATTRIBUTES * TILESET_DIM_ATTRIBUTES;
-
-static_assert(TILESET_DIM == (1 << TILESET_DIM_LOG2));
-
-enum TilesetTileType : s32 {
-	TILE_EMPTY = 0,
-	TILE_SOLID = 1,
-	TILE_TYPE_COUNT
-};
-
-#ifdef EDITOR
-constexpr const char* METATILE_TYPE_NAMES[TILE_TYPE_COUNT] = { "Empty", "Solid" };
-#endif
-
-struct TilesetTile {
-	s32 type;
-	Metatile metatile;
-};
-
-struct Tileset {
-	TilesetTile tiles[TILESET_SIZE];
-};
-
-struct Tilemap {
-	u32 width;
-	u32 height;
-	TilesetHandle tilesetHandle;
-	u32 tilesOffset;
-
-	inline u8* GetTileData() const {
-		return (u8*)this + tilesOffset;
-	}
-};
