+++ src/spawner.cpp
@@ -1,6 +1,4 @@
-#include "spawner.h"
 #include "actors.h"
-#include "actor_prototype_types.h"
 #include "random.h"
 #include "asset_manager.h"
