--- src/spawner.h
@@ -1,8 +1,0 @@
-#pragma once
-#include "spawner_data.h"
-
-namespace Game {
-	extern const ActorInitFn spawnerInitTable[SPAWNER_TYPE_COUNT];
-	extern const ActorUpdateFn spawnerUpdateTable[SPAWNER_TYPE_COUNT];
-	extern const ActorDrawFn spawnerDrawTable[SPAWNER_TYPE_COUNT];
-}
